refactor : Like -> CommentLike 로 이름 변경
- 포스트라이크와의 혼동을 피하기 위해서 엔티티 필드명을 변경했습니다.

refactor : 다른 서비스 Repository 의존을 Service 의존으로 변경

- 타 도메인 Repository 를 의존하는 것은 좋지 않습니다. 여러 이유가 있습니다.
- 이 경우 코멘트 서비스가 포스트 서비스에 존재하는 것이 자연스러울 수 있습니다. 도메인간 의존관계를 어느 쪽이 컨트롤하냐를 고민해봐야 할 것 같아요. 다른 해결 방법으로는 새로운 파사드를 선언하는 방법도 있습니다.

refactor : 사용하지 않는 필드, 주석 제거

- 사용하지 않는 필드는 제거했습니다. 필요 시 깃 로그로 복원하는것이 좋겠습니다.

+++ src/main/java/com/example/yeogiserver/comment/application/CommentService.java
@@ -8,8 +8,8 @@ import com.example.yeogiserver.comment.domain.CommentRepository;
 import com.example.yeogiserver.event.recommand.RecommandEvent;
 import com.example.yeogiserver.member.application.MemberQueryService;
 import com.example.yeogiserver.member.domain.Member;
+import com.example.yeogiserver.post.application.PostService;
 import com.example.yeogiserver.post.domain.Post;
-import com.example.yeogiserver.post.domain.PostRepository;
 import com.example.yeogiserver.security.domain.CustomUserDetails;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
@@ -25,8 +25,8 @@ import java.util.List;
 public class CommentService {
     private final CommentRepository commentRepository;
     private final LikeService likeService;
-    private final PostRepository postRepository;
-    private final MemberQueryService memberRepository;
+    private final PostService postService;
+    private final MemberQueryService memberQueryService;
     private final ApplicationEventPublisher applicationEventPublisher;
 
     public List<CommentResponseDto> getComments(Long postId, Long memberId, Pageable pageable) {
@@ -50,15 +50,15 @@ public class CommentService {
 
     public CommentSaveResponse addComment(CommentRequestDto commentRequestDto, CustomUserDetails userDetails) {
 
-        Member member = memberRepository.findMember(userDetails.getEmail());
-        Post post = postRepository.findById(commentRequestDto.postId()).orElseThrow(() -> new IllegalArgumentException("Post not found"));
+        Member member = memberQueryService.findMember(userDetails.getEmail());
+        Post post = postService.getPost(commentRequestDto.postId());
         applicationEventPublisher.publishEvent(new RecommandEvent(this,post.getId(),member.getId()));
 
         return CommentSaveResponse.of(commentRepository.saveComment(Comment.of(member,commentRequestDto.content(),post)));
     }
     public CommentSaveResponse addReply(CommentRequestDto commentRequestDto,CustomUserDetails userDetails, Long commentId) {
-        Member member = memberRepository.findMember(userDetails.getEmail());
-        Post post = postRepository.findById(commentRequestDto.postId()).orElseThrow(() -> new IllegalArgumentException("Post not found"));
+        Member member = memberQueryService.findMember(userDetails.getEmail());
+        Post post = postService.getPost(commentRequestDto.postId());
 
         Comment comment = commentRepository.findById(commentId).orElseThrow(()-> new RuntimeException("Comment Not Found"));
         Comment child = Comment.of(member,commentRequestDto.content(),post);
@@ -67,6 +67,7 @@ public class CommentService {
         commentRepository.saveComment(child);
         return CommentSaveResponse.of(child);
     }
+
     public CommentSaveResponse updateComment(Long id, CommentRequestDto commentRequestDto) {
 
         Comment comment = commentRepository.findById(id)
@@ -75,17 +76,16 @@ public class CommentService {
         comment.update(commentRequestDto.content());
         return CommentSaveResponse.of(comment);
     }
-    public void deleteComment(Long id) {
-        //TODO. Session and Request Validate
+
+    public void deleteComment(Long id, Long memberId) {
         Comment comment = commentRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Comment not found"));
 
-        commentRepository.delete(comment);
-    }
-    public void deletePostComment(Long postId) {
-        //TODO. Session and Request Validate
+        if (!comment.getMember().getId().equals(memberId)){
+            throw new IllegalArgumentException("Not my Comment");
+        }
 
-        commentRepository.deleteByPostId(postId);
+        commentRepository.delete(comment);
     }
 
     public Long getCommentCount(Long postId) {
